Add files via upload
NYT article on: 

Ozempic (and other GLP-1s) are an amazing product. It makes you lose 15-25% predictably based on drug and dose. 

However it offers a tradeoff: quick weight loss, but often also muscle loss. 

To counteract this tradeoff, companies are pushing drugs and services to retain/increase muscle mass. Is there increased demand after a patient receives GLP-1 therapy for more fitness/nutritional intervention as a result of this tradeoff? Will that accelerate customer demand for fitness services?